Changes according to code review Added javadoc plugin execution
